Clear chapter dataLoader cache in mutations
E.g. in case the global update is running which fetches new chapters for a manga and after this, while the update is still running, a chapter of this manga gets updated, the dataloader will still contain the outdated chapter state from before the mutation.

In case of how the updater subscription works, this will cause the subscription to send the outdated chapter state to the client, which then overwrites the correct chapter state with the outdated one from the subscription